requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32
flask_sqlalchemy==3.1.1
pandas==2.0.3
openpyxl==3.1.5
xlrd==2.0.1
